+++ app/build.gradle
@@ -4,6 +4,7 @@ plugins {
     id 'org.jetbrains.kotlin.plugin.serialization'
     id 'kotlin-parcelize'
     id 'kotlin-kapt'
+
 }
 
 android {
@@ -66,4 +67,7 @@ dependencies {
     implementation("com.squareup.okhttp3:logging-interceptor:4.9.0")
     implementation "androidx.viewpager2:viewpager2:1.0.0"
     annotationProcessor 'androidx.room:room-compiler:2.2.6'
+    implementation "org.koin:koin-android:2.2.2"
+    implementation "org.koin:koin-android-scope:2.2.2"
+    implementation "org.koin:koin-android-viewmodel:2.2.2"
 }
