+++ app/src/main/AndroidManifest.xml
@@ -13,11 +13,11 @@
         android:theme="@style/Theme.KotlinOne">
 
         <activity
-            android:name=".View.FragmentMovieDetails"
+            android:name=".movieDetails.FragmentMovieDetails"
             tools:ignore="Instantiatable" />
         <activity android:name=".MovieDetailsActivity"
             tools:ignore="MissingClass" />
-        <activity android:name=".View.MainActivity">
+        <activity android:name=".MainActivity">
 
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
