+++ app/src/main/java/ru/firstSet/kotlinOne/MainActivity.kt
@@ -1,10 +1,9 @@
-package ru.firstSet.kotlinOne.View
+package ru.firstSet.kotlinOne
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
-import ru.firstSet.kotlinOne.Data.MovieRepository
-import ru.firstSet.kotlinOne.R
-import ru.firstSet.kotlinOne.viewModel.ViewModelMoviesList
+import ru.firstSet.kotlinOne.movieList.ViewModelMoviesList
+import ru.firstSet.kotlinOne.movieList.FragmentMoviesList
 
 class MainActivity : AppCompatActivity() {
 
@@ -20,8 +19,6 @@ class MainActivity : AppCompatActivity() {
     companion object {
         const val FRAGMENT_TAG_MOVIES_LIST = "MoviesList"
         const val FRAGMENT_TAG_MOVIES_DETAILS = "MoviesDetails"
-        val viewModel:ViewModelMoviesList= ViewModelMoviesList
-
     }
 
     fun callFragmentMoviesList() {
