+++ app/src/main/java/ru/firstSet/kotlinOne/data/DBContract.kt
@@ -1,14 +1,9 @@
-package ru.firstSet.kotlinOne.Data
+package ru.firstSet.kotlinOne.data
 
 object DBContract {
     const val DATABASE_NAME = "moviesDB2.db"
     public object MovieColumns {
-        const val TABLE_NAME_NOWPLAYING = "now_playing"
-        const val TABLE_NAME_Popular = "popular"
-        const val TABLE_NAME_TopRated = "top_rated"
-        const val TABLE_NAME_UpComing= "upcoming"
         const val TABLE_NAME = "moviesTable"
-
         const val COLUMN_NAME_ID = "id"
         const val COLUMN_NAME_TITLE = "title"
         const val COLUMN_NAME_POSTERPICTURE = "posterPicture"
@@ -20,13 +15,11 @@ object DBContract {
         const val COLUMN_NAME_VOTE_COUNT = "vote_count"
         const val COLUMN_NAME_GENREIDS = "genres"
         const val COLUMN_NAME_ACTORS = "actors"
-
         const val TABLE_NAME_ACTOR = "actorTable"
         const val COLUMN_NAME_ACTOR_MOVIEID = "actorMovieId"
         const val COLUMN_NAME_ACTOR_ID = "actorId"
         const val COLUMN_NAME_ACTOR_PICTURE = "picture"
         const val COLUMN_NAME_ACTOR_NAME = "actorName"
-
         const val TABLE_NAME_GENRE = "genreTable"
         const val COLUMN_NAME_GENRE_MOVIEID = "genreMovieId"
         const val COLUMN_NAME_GENRE_ID = "genreId"
