+++ app/src/main/java/ru/firstSet/kotlinOne/movieDetails/ActorsAdapter.kt
@@ -1,4 +1,4 @@
-package ru.firstSet.kotlinOne.View
+package ru.firstSet.kotlinOne.movieDetails
 
 import android.view.LayoutInflater
 import android.view.View
@@ -7,7 +7,7 @@ import android.widget.ImageView
 import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
-import ru.firstSet.kotlinOne.Data.ActorEntity
+import ru.firstSet.kotlinOne.data.ActorEntity
 import ru.firstSet.kotlinOne.R
 
 class ActorsAdapter(val actorList: List<ActorEntity>) :
