+++ app/src/main/java/ru/firstSet/kotlinOne/repository/MovieApi.kt
@@ -1,13 +1,5 @@
-package ru.firstSet.kotlinOne.Data
+package ru.firstSet.kotlinOne.data
 
-import com.jakewharton.retrofit2.converter.kotlinx.serialization.asConverterFactory
-import kotlinx.serialization.json.Json
-import okhttp3.Interceptor
-import okhttp3.MediaType.Companion.toMediaType
-import okhttp3.OkHttpClient
-import okhttp3.Response
-import okhttp3.logging.HttpLoggingInterceptor
-import retrofit2.Retrofit
 import retrofit2.http.GET
 import retrofit2.http.Path
 import ru.firstSet.kotlinOne.ResultGenre
