+++ app/src/main/java/ru/firstSet/kotlinOne/repository/Repositorys.kt
@@ -1,16 +1,16 @@
-package ru.firstSet.kotlinOne.Data
+package ru.firstSet.kotlinOne.repository
 
-import android.content.Context
-import android.util.Log
 import kotlinx.coroutines.*
-import ru.firstSet.kotlinOne.Data.MovieDatabase.Companion.createMovieDatabaseInstance
+import ru.firstSet.kotlinOne.GenreEntity
+import ru.firstSet.kotlinOne.data.ActorEntity
+import ru.firstSet.kotlinOne.data.Movie
+import ru.firstSet.kotlinOne.data.MovieEntity
+import ru.firstSet.kotlinOne.data.SeachMovie
+import ru.firstSet.kotlinOne.repository.RemoteDataStore.Companion.BASE_URL_MOVIES
 import java.util.*
 
 @Suppress("NULLABILITY_MISMATCH_BASED_ON_JAVA_ANNOTATIONS")
-class MovieRepository(context: Context) {
-    private val BASE_URL_MOVIES = "https://image.tmdb.org/t/p/original"
-    private val remoteDataStore = ru.firstSet.kotlinOne.Data.RemoteDataStore
-    val movieDataBase: MovieDatabase = createMovieDatabaseInstance(context)
+class Repositorys(val remoteDataStore : RemoteDataStore, val movieDataBase : MovieDatabase) {
 
     suspend fun loadGenreFromNET(idMovie: Long): List<GenreEntity> =
         remoteDataStore.getGenreFromNet().genres.map {
@@ -20,7 +20,6 @@ class MovieRepository(context: Context) {
                 genreMovieId = idMovie
             )
         }
-
 
     suspend fun loadRuntimesFromNET(id: Long): Int = withContext(Dispatchers.IO) {
         remoteDataStore.getSearchRuntimes(id).runtime
@@ -35,18 +34,6 @@ class MovieRepository(context: Context) {
                 actorMovieId = movieId
             )
         }
-    }
-
-    suspend fun readActorFromDb(idMovie: Long): List<ActorEntity> {
-        var actorList: List<ActorEntity> = listOf()
-        withContext(Dispatchers.IO) {
-            actorList = movieDataBase.movieDAO.getActorFromSQL(idMovie)
-            Log.v("readActorFromDb ", " ${actorList.size}")
-            actorList.forEach() {
-                Log.v("readActorFromDb ", " ${it.actorMovieId},  ${it.actorName}, ${it.actorId}")
-            }
-        }
-        return actorList
     }
 
     suspend fun loadMoviesFromNET(seachMovie: String): List<Movie> = withContext(Dispatchers.IO) {
@@ -71,44 +58,27 @@ class MovieRepository(context: Context) {
         }
     }
 
+    suspend fun readActorFromDb(idMovie: Long): List<ActorEntity> =
+        withContext(Dispatchers.IO) {
+            movieDataBase.movieDAO.getActorFromSQL(idMovie)
+        }
 
     suspend fun readMovieFromDb(seachMovie: SeachMovie): List<Movie> =
         withContext(Dispatchers.IO) {
-            Log.v("readMovieFromDb  ", " ${seachMovie.seachMovie}")
-
-            val moviesEntity: List<MovieEntity> = movieDataBase.movieDAO.getAllMovies(seachMovie.seachMovie)
+            val moviesEntity: List<MovieEntity> =
+                movieDataBase.movieDAO.getAllMovies(seachMovie.seachMovie)
             convertMovieEntityToMovie(moviesEntity)
         }
 
-    suspend fun readGenreFromDb(idMovie: Long): List<GenreEntity> {
-        var listGenre: List<GenreEntity> = mutableListOf()
+    suspend fun readGenreFromDb(idMovie: Long): List<GenreEntity> =
         withContext(Dispatchers.IO) {
-            Log.v("idMovie =  ", " ${idMovie}")
-            listGenre = movieDataBase.movieDAO.getGenresFromSQL(idMovie)
-
-        }
-//        listGenre.forEach(){
-            Log.v("listGenreFromDb ", " ${listGenre.size}")
-//        }
-        readAllGenres()
-        return listGenre
-    }
-
-    suspend fun readAllGenres(): List<GenreEntity> {
-        var listGenre: List<GenreEntity> = listOf()
-        withContext(Dispatchers.IO) {
-            listGenre = movieDataBase.movieDAO.getAllGenres()
-            Log.v("readAllGenres", " ${listGenre.size}")
+            movieDataBase.movieDAO.getGenresFromSQL(idMovie)
         }
 
-        listGenre.forEach() {
-            Log.v("list ", " ${it.genreMovieId},  ${it.name}, ${it.idGenre}")
-        }
-        return listGenre
-    }
-
-
-    private fun convertMovieToMovieEntity(movieList: List<Movie>,seachMovie:SeachMovie): List<MovieEntity> {
+    private fun convertMovieToMovieEntity(
+        movieList: List<Movie>,
+        seachMovie: SeachMovie
+    ): List<MovieEntity> {
         val newMovie: MutableList<MovieEntity> = mutableListOf()
         movieList.forEach() {
             val movie: MovieEntity = MovieEntity(
@@ -121,14 +91,12 @@ class MovieRepository(context: Context) {
                 overview = it.overview,
                 adult = it.adult,
                 vote_count = it.vote_count,
-                seachMovie=seachMovie.seachMovie
-
+                seachMovie = seachMovie.seachMovie
             )
             newMovie.add(movie)
         }
         return newMovie
     }
-
 
     private suspend fun convertMovieEntityToMovie(movieEntity: List<MovieEntity>): List<Movie> {
         val newMovie: MutableList<Movie> = mutableListOf()
@@ -147,38 +115,22 @@ class MovieRepository(context: Context) {
                 actors = listOf()
             )
             newMovie.add(movie)
-            Log.v("convertMovieEntity", " ${movie.genres.size}")
-            Log.v("convertMovieEntity", " ${movie.title}")
-
         }
         return newMovie
     }
 
-    suspend fun saveMovieToDB(movies: List<Movie>,seachMovie:SeachMovie) {
-        return withContext(Dispatchers.IO) {
-            Log.v("saveMovieToDB (movies)", " ${movies.size}")
-//            if (movies.size > 0) {
-//                movieDataBase.movieDAO.deleteAllMovies()
-//            }
-            movieDataBase.movieDAO.saveMovies(convertMovieToMovieEntity(movies,seachMovie))
+    suspend fun saveMovieToDB(movies: List<Movie>, seachMovie: SeachMovie) {
+        withContext(Dispatchers.IO) {
+            movieDataBase.movieDAO.saveMovies(convertMovieToMovieEntity(movies, seachMovie))
             movies.forEach() {
                 movieDataBase.movieDAO.saveGenres(it.genres)
-                Log.v("insertGenres", " ${it.genres.size}")
             }
         }
     }
 
     suspend fun saveActorToDB(actors: List<ActorEntity>) {
-        return withContext(Dispatchers.IO) {
-            Log.v("saveActorToDB", " ${actors.size}")
-            if (actors.size > 0) {
-                movieDataBase.movieDAO.deleteActors()
-            }
+        withContext(Dispatchers.IO) {
             movieDataBase.movieDAO.saveActors(actors)
         }
     }
-
-
 }
-
-
